+++ lib/app/pages/home_page/widgets/month_resume_widget/month_resume_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:personal_expense_tracker/app/providers/transaction_provider.dart';
+import 'package:personal_expense_tracker/app/providers/transaction_provider/transaction_cache.dart';
 import 'package:personal_expense_tracker/app/widgets/card_widget/card_widget.dart';
 import 'package:personal_expense_tracker/app/widgets/graph_bar_widget/graph_bar_widget.dart';
